+++ internal/controller/postgres_controller_test.go
@@ -6,7 +6,7 @@ import (
 	"github.com/nais/pgrator/internal/config"
 	"github.com/nais/pgrator/internal/synchronizer"
 	"github.com/nais/pgrator/internal/synchronizer/ownership"
-	iam_cnrm_cloud_google_com_v1beta1 "github.com/nais/pgrator/internal/thirdparty/google/v1beta1"
+	"github.com/nais/pgrator/internal/thirdparty/google/iam/v1beta1"
 	data_nais_io_v1 "github.com/nais/pgrator/pkg/api/datav1"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
